# packages/api-schema/openapi.yaml
openapi: 3.1.0
info:
  title: WISE Workbench API
  description: 'WISE Workbench backend: projects, datasets, mappings, case tables, norm versions, runs, backlogs, slice detail, traces, diagnostics, signal distributions, process flow and jobs. Errors are RFC 9457 problem details.'
  version: 0.1.0
paths:
...
  /projects/{projectId}/norms:
    get:
      tags:
      - norms
      summary: List Norms
      operationId: listNorms
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NormVersion'
                title: Response Listnorms
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - norms
      summary: Create Norm Version
      description: Body is the library's norm JSON plus a note; a new immutable version is created and validated.
      operationId: createNormVersion
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormVersionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormVersion'
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unprocessable Content
  /projects/{projectId}/norms/{normVersionId}:
    get:
      tags:
      - norms
      summary: Get Norm Version
      description: A norm version; `warnings` lists activities and attributes it names that never occur in the case table (recomputed against `caseTableId` when given).
      operationId: getNormVersion
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: normVersionId
        in: path
        required: true
        schema:
          type: string
          title: Normversionid
      - name: caseTableId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Casetableid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormVersion'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - norms
      summary: Set Norm Status
      operationId: setNormStatus
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: normVersionId
        in: path
        required: true
        schema:
          type: string
          title: Normversionid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormVersion'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/norms/{normVersionId}/check:
    post:
      tags:
      - norms
      summary: Check Norm
      operationId: checkNorm
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: normVersionId
        in: path
        required: true
        schema:
          type: string
          title: Normversionid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormCheckRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormCheck'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /projects/{projectId}/notebook:
    get:
      tags:
      - notebook
      summary: Get Notebook
      operationId: getNotebook
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notebook'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/notebook/snapshots:
    post:
      tags:
      - notebook
      summary: Create Snapshot
      description: 'Freeze a screen: multipart with an optional PNG `image` and a JSON `payload` ({title, note, context {run_id, slicing, view, filters, url, screen}, data, author}); the title may also come as a form field.'
      operationId: createSnapshot
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_createSnapshot'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/notebook/reorder:
    post:
      tags:
      - notebook
      summary: Reorder Snapshots
      description: The snapshots in the given order; snapshots not listed keep their relative order after them.
      operationId: reorderSnapshots
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotOrder'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notebook'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/notebook/export:
    get:
      tags:
      - notebook
      summary: Export Notebook
      description: 'The notebook as a zip: `notebook.md` with the images under `images/` (format=markdown); PowerPoint arrives in cycle 4.'
      operationId: exportNotebook
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: format
        in: query
        required: false
        schema:
          enum:
          - markdown
          - pptx
          type: string
          default: markdown
          title: Format
      responses:
        '200':
          description: a zip file
          content:
            application/json:
              schema: {}
            application/zip: {}
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unprocessable Content
  /projects/{projectId}/notebook/snapshots/{snapshotId}:
    get:
      tags:
      - notebook
      summary: Get Snapshot
      operationId: getSnapshot
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: snapshotId
        in: path
        required: true
        schema:
          type: string
          title: Snapshotid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - notebook
      summary: Update Snapshot
      operationId: updateSnapshot
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: snapshotId
        in: path
        required: true
        schema:
          type: string
          title: Snapshotid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - notebook
      summary: Delete Snapshot
      operationId: deleteSnapshot
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: snapshotId
        in: path
        required: true
        schema:
          type: string
          title: Snapshotid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/notebook/snapshots/{snapshotId}/image:
    get:
      tags:
      - notebook
      summary: Get Snapshot Image
      operationId: getSnapshotImage
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: snapshotId
        in: path
        required: true
        schema:
          type: string
          title: Snapshotid
      responses:
        '200':
          description: the PNG
          content:
            application/json:
              schema: {}
            image/png: {}
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - notebook
      summary: Replace Snapshot Image
      description: Replace the snapshot's PNG.
      operationId: replaceSnapshotImage
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: snapshotId
        in: path
        required: true
        schema:
          type: string
          title: Snapshotid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_replaceSnapshotImage'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_createSnapshot:
      properties:
        image:
          anyOf:
          - type: string
            contentMediaType: application/octet-stream
          - type: 'null'
          title: Image
        payload:
          anyOf:
          - type: string
          - type: 'null'
          title: Payload
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      title: Body_createSnapshot
    Body_replaceSnapshotImage:
      properties:
        image:
          type: string
          contentMediaType: application/octet-stream
          title: Image
      type: object
      required:
      - image
      title: Body_replaceSnapshotImage
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NormCheck:
      properties:
        constraints:
          items:
            $ref: '#/components/schemas/NormCheckConstraint'
          type: array
          title: Constraints
        issues:
          items:
            type: string
          type: array
          title: Issues
        warnings:
          items:
            type: string
          type: array
          title: Warnings
        cases:
          anyOf:
          - type: integer
          - type: 'null'
          title: Cases
        fingerprint:
          anyOf:
          - type: string
          - type: 'null'
          title: Fingerprint
      type: object
      required:
      - constraints
      title: NormCheck
    NormCheckConstraint:
      properties:
        id:
          type: string
          title: Id
        layer:
          anyOf:
          - type: string
          - type: 'null'
          title: Layer
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        activitiesMissing:
          items:
            type: string
          type: array
          title: Activitiesmissing
        casesInScope:
          type: integer
          title: Casesinscope
        casesEvaluated:
          type: integer
          title: Casesevaluated
      type: object
      required:
      - id
      - casesInScope
      - casesEvaluated
      title: NormCheckConstraint
    NormCheckRequest:
      properties:
        caseTableId:
          type: string
          title: Casetableid
      type: object
      required:
      - caseTableId
      title: NormCheckRequest
    NormStatusUpdate:
      properties:
        status:
          type: string
          enum:
          - draft
          - reviewed
          - approved
          title: Status
      type: object
      required:
      - status
      title: NormStatusUpdate
    NormVersion:
      properties:
        id:
          type: string
          title: Id
        version:
          type: integer
          title: Version
        fingerprint:
          type: string
          title: Fingerprint
        status:
          type: string
          enum:
          - draft
          - reviewed
          - approved
          title: Status
        note:
          type: string
          title: Note
          default: ''
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        parentId:
          anyOf:
          - type: string
          - type: 'null'
          title: Parentid
        norm:
          additionalProperties: true
          type: object
          title: Norm
        validation:
          items:
            type: string
          type: array
          title: Validation
        warnings:
          items:
            type: string
          type: array
          title: Warnings
          description: 'Norm.check: activities or attributes the norm names that never occur'
        guidance_complete:
          type: boolean
          title: Guidance Complete
          description: every layer and expectation has a plain name
          default: false
        createdAt:
          type: string
          format: date-time
          title: Createdat
        normId:
          type: string
          title: Normid
        name:
          type: string
          title: Name
        views:
          items:
            type: string
          type: array
          title: Views
      type: object
      required:
      - id
      - version
      - fingerprint
      - status
      - norm
      - createdAt
      - normId
      - name
      title: NormVersion
    NormVersionCreate:
      properties:
        norm:
          additionalProperties: true
          type: object
          title: Norm
          description: library norm JSON
        note:
          type: string
          title: Note
        parentId:
          anyOf:
          - type: string
          - type: 'null'
          title: Parentid
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
      type: object
      required:
      - norm
      - note
      title: NormVersionCreate
    Notebook:
      properties:
        projectId:
          type: string
          title: Projectid
        snapshots:
          items:
            $ref: '#/components/schemas/Snapshot'
          type: array
          title: Snapshots
        exportFormats:
          items:
            type: string
          type: array
          title: Exportformats
      type: object
      required:
      - projectId
      - snapshots
      title: Notebook
...
    Problem:
      properties:
        type:
          type: string
          title: Type
          default: about:blank
        title:
          type: string
          title: Title
        status:
          type: integer
          title: Status
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
        instance:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance
        code:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
        errors:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Errors
      type: object
      required:
      - title
      - status
      title: Problem
...
    Snapshot:
      properties:
        id:
          type: string
          title: Id
        projectId:
          type: string
          title: Projectid
        title:
          type: string
          title: Title
        note:
          type: string
          title: Note
          default: ''
        context:
          additionalProperties: true
          type: object
          title: Context
        data:
          title: Data
        hasImage:
          type: boolean
          title: Hasimage
          default: false
        imageUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Imageurl
        order:
          type: integer
          title: Order
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - projectId
      - title
      - order
      - createdAt
      - updatedAt
      title: Snapshot
    SnapshotOrder:
      properties:
        ids:
          items:
            type: string
          type: array
          title: Ids
      type: object
      required:
      - ids
      title: SnapshotOrder
    SnapshotUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        context:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Context
        data:
          title: Data
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
      type: object
      title: SnapshotUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
servers:
- url: /api/v1